a rough script to build a CP genome from our data
this is a first attempt, but so far all I’m really doing is:

1. Mapping everything to the reference with BWA
2. Keeping the reads that map to the chloroplast
3. Using those to run CANU

An obvious drawback is that my chloroplast in the reference is a single
circular molecular that’s been linearised. So, I might have an
ascertainment bias where I miss reads that span the artificial cut in
the circular genome. A simple fix for this is to just double up the CP
genome in the reference (literally a copy-paste should work) so reads
that span the break have somewhere sensible to map.

For working without a reference CP genome, we could do something
similar and just extract reads that map (or just BLAST match) to any of
a range of CP genomes from related taxa. Various options for doing
that, all are probably fine.